pyproject.toml
[project]
name = "airflow3-starter"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=3.1.6",
    "aiohttp>=3.9.0",
]
